refactor(core-bridge-adapter-sdk): rename BridgeSdk file to BridgeAdapterSdk refactor(core-bridge-adapter-sdk): rename AbstractBridgeAdapter's getQuoteDetail to getSwapDetail to be more aligned with the public Api
