+++ packages/elasticbottle-core-bridge-adapter-sdk/src/index.ts
@@ -1,6 +1,6 @@
 export { CHAIN_NAMES as SupportedChainNames } from "./constants/ChainNames";
-export { BridgeAdapterSdk } from "./lib/BridgeSdk";
-export type { BridgeAdapterSdkArgs } from "./lib/BridgeSdk";
+export { BridgeAdapterSdk } from "./lib/BridgeAdapterSdk";
+export type { BridgeAdapterSdkArgs } from "./lib/BridgeAdapterSdk";
 export type { BridgeStatus, SolanaOrEvmAccount } from "./types/Bridges";
 export type { ChainName, ChainSourceAndTarget } from "./types/Chain";
 export type { ChainDestType } from "./types/ChainDest";
