+++ packages/elasticbottle-core-bridge-adapter-sdk/src/lib/BridgeAdapter/DeBridgeBridgeAdapter.ts
@@ -245,7 +245,7 @@ export class DeBridgeBridgeAdapter extends AbstractBridgeAdapter {
     return this.tokenList[chain];
   }
 
-  async getQuoteDetails(
+  async getSwapDetails(
     sourceToken: TokenWithAmount,
     targetToken: Token
   ): Promise<SwapInformation> {
